+++ tests/integration/server-hook/middleware/server/_middleware.ts
@@ -2,13 +2,20 @@ import { Middleware } from '@modern-js/runtime/server';
 
 export const middleware: Middleware[] = [
   (ctx, next) => {
-    const { res } = ctx.source;
-    res.setHeader('x-unstable-middleware', 'true');
+    if (ctx.request.url === '/') {
+      const { res } = ctx.source;
+      res.setHeader('x-unstable-middleware', 'true');
+    }
+
     next();
   },
   (ctx, next) => {
-    const { res } = ctx.source;
-    res.end('hello modern');
+    if (ctx.request.url === '/') {
+      const { res } = ctx.source;
+
+      res.end('hello modern');
+    }
+
     next();
   },
 ];