+++ tests/integration/server-hook/new-middleware/shared/index.ts
@@ -1,0 +1,7 @@
+interface User {
+  name: string;
+}
+
+export type Var = {
+  user: User;
+};