+++ packages/cli/plugin-ssg/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/plugin-ssg
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/utils@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes