+++ packages/server/plugin-koa/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-koa
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/bff-core@2.49.4
+- @modern-js/bff-runtime@2.49.4
+- @modern-js/types@2.49.4
+- @modern-js/utils@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes