+++ packages/toolkit/node-bundle-require/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/node-bundle-require
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/utils@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes